+++ src/bot/commands/ticket/open.ts
@@ -1,7 +1,6 @@
 import { Command } from '../../commandHandler.js';
 import { Message } from 'discord.js';
 import { settings } from '../../bot.js';
-import { getTicket } from '../../lib/searchMessage.js';
 import { sendSuccess, sendError, sendInfo } from '../../lib/embeds.js';
 import log from '../../lib/log.js';
 import { openTicket } from '../../lib/tickets.js';
@@ -14,19 +13,19 @@ export const command: Command = {
     maxArgs: null,
     superCommands: ['ticket'],
     callback: async (message: Message, args: string[], text: string) => {
-        const { guild, channel } = message;
-        if (!guild || channel.id !== settings.ticket.managementChannelID) return;
+        const { guild, channel, member } = message;
+        if (!guild || !member || channel.id !== settings.ticket.managementChannelID) return;
+
+        const loadingEmbed = await sendInfo(channel, 'Opening ticket...');
 
         try {
-            let ticket = await getTicket(message, args, text, true);
-
-            const loadingEmbed = await sendInfo(channel, 'Opening ticket...');
-            await openTicket(ticket, guild);
+            const { title } = await openTicket(args, text, member);
             await loadingEmbed.delete();
 
             sendSuccess(channel, 'Ticket opened.');
-            log(`<@${message.author.id}> opened the ticket "${ticket.title}".`);
+            log(`<@${message.author.id}> opened the ticket "${title}".`);
         } catch (error) {
+            await loadingEmbed.delete();
             if (error) sendError(channel, error);
         }
     },
